db-auths: Refactored datasets tests in separate files. Adjusted kerberos approach on db_connector
