+++ webserver/tests/test_beacon_datasets.py
@@ -1,0 +1,87 @@
+from sqlalchemy.exc import ProgrammingError, OperationalError
+from unittest.mock import Mock
+
+
+class TestBeacon:
+    def test_beacon_available_to_admin(
+            self,
+            client,
+            post_json_admin_header,
+            mocker,
+            dataset
+    ):
+        """
+        Test that the beacon endpoint is accessible to admin users
+        """
+        mocker.patch('app.helpers.query_validator.create_engine')
+        mocker.patch(
+            'app.helpers.query_validator.sessionmaker',
+        ).__enter__.return_value = Mock()
+        response = client.post(
+            "/datasets/selection/beacon",
+            json={
+                "query": "SELECT * FROM table_name",
+                "dataset_id": dataset.id
+            },
+            headers=post_json_admin_header
+        )
+        assert response.status_code == 200
+        assert response.json['result'] == 'Ok'
+
+    def test_beacon_available_to_admin_invalid_query(
+            self,
+            client,
+            post_json_admin_header,
+            mocker,
+            dataset
+    ):
+        """
+        Test that the beacon endpoint is accessible to admin users
+        """
+        mocker.patch('app.helpers.query_validator.create_engine')
+        mocker.patch(
+            'app.helpers.query_validator.sessionmaker',
+            side_effect = ProgrammingError(statement="", params={}, orig="error test")
+        )
+        response = client.post(
+            "/datasets/selection/beacon",
+            json={
+                "query": "SELECT * FROM table",
+                "dataset_id": dataset.id
+            },
+            headers=post_json_admin_header
+        )
+        assert response.status_code == 500
+        assert response.json['result'] == 'Invalid'
+
+    def test_beacon_connection_failed(
+            self,
+            client,
+            post_json_admin_header,
+            mocker,
+            dataset
+    ):
+        """
+        Test that the beacon endpoint is accessible to admin users
+        but returns an appropriate error message in case of connection
+        failed
+        """
+        mocker.patch('app.helpers.query_validator.create_engine')
+        mocker.patch(
+            'app.helpers.query_validator.sessionmaker',
+            side_effect = OperationalError(
+                statement="Unable to connect: Adaptive Server is unavailable or does not exist",
+                params={}, orig="error test"
+            )
+        )
+        response = client.post(
+            "/datasets/selection/beacon",
+            json={
+                "query": "SELECT * FROM table",
+                "dataset_id": dataset.id
+            },
+            headers=post_json_admin_header
+        )
+        assert response.status_code == 500
+        assert response.json['error'] == 'Could not connect to the database'
+
